requirement.txt
requests
urllib3
pandas
python-dotenv